+++ pcsx2/pcsx2core.vcxproj.filters
@@ -1606,7 +1606,7 @@
     <ClInclude Include="DEV9\ATA\HddCreate.h">
       <Filter>System\Ps2\DEV9\ATA</Filter>
     </ClInclude>
-    <ClInclude Include="DEV9\Config.h">
+    <ClInclude Include="DEV9\DEV9Config.h">
       <Filter>System\Ps2\DEV9</Filter>
     </ClInclude>
     <ClInclude Include="DEV9\DEV9.h">
